builder class testing (#7)
* builder class testing

* deleting old data class from w/end tests

* moved a couple of helper functions out of data builder and into helpers

* fixed an issue with the imports and a few other var naming issues
